requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
motor==3.0.0
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
zipp==3.9.0
